samples: Bluetooth: Refactor broadcaster multiple for bsim test reuse
Refactor the broadcaster multiple sample so that it can be
reused as bsim test to verify Extended advertisement
chaining.
