+++ samples/bluetooth/broadcaster_multiple/CMakeLists.txt
@@ -4,4 +4,7 @@ cmake_minimum_required(VERSION 3.20.0)
 find_package(Zephyr REQUIRED HINTS $ENV{ZEPHYR_BASE})
 project(broadcaster_multiple)
 
-target_sources(app PRIVATE src/main.c)
+target_sources(app PRIVATE
+  src/main.c
+  src/broadcaster_multiple.c
+)
